perf: provide defaults directly in `Args` struct
This commit provides reasonable defaults in the `Args` struct directly.
Previously, we used a `NormalizedArgs` struct from which we accessed
normalized arguments and where we transformed the parsed cli arguments
to non-optional fields.
This isn't needed since we can simply provide defaults in the `Args`
struct directly.

Additionally, functions that used to live in `NormalizedArgs` struct
are now implemeneted in the `Args` struct.

Resolves #2